requirements.txt
gradio
livekit-agents
livekit-plugins-silero
livekit-plugins-openai
openai
dashscope
python-dotenv
numpy
scipy
